# Conflicts:
#	archetype-examples/src/main/resources/archetype-resources/src/main/java/validator/ResponseFromFileValidator.java
#	archetype-examples/src/main/resources/archetype-resources/src/main/resources/suite/listeners.conf.xml
#	archetype-examples/src/main/resources/archetype-resources/src/main/resources/suite/suite.conf.xml
#	archetype-examples/src/main/resources/archetype-resources/src/main/resources/suite/test.description.conf.xml
#	archetype-java-builders/src/main/resources/archetype-resources/src/main/java/ExampleJLoadScenarioProvider.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/collector/NotNullResponseValidator.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/collector/ResponseValidator.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/collector/invocation/InvocationInfo.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/collector/invocation/InvocationListener.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/scenario/TerminateByDuration.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/scenario/TerminateByTotalSampling.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/services/DefaultMetricService.java
#	chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/services/MetricService.java
#	chassis/core/src/main/java/com/griddynamics/jagger/invoker/Invoker.java
#	chassis/core/src/test/java/com/griddynamics/jagger/ExampleJLoadScenarioProvider.java
#	doc/css/my_customdoxygen.css
#	doc/pages/ListenersMain.dox
#	doc/pages/MetricsCustom.dox
#	doc/pages/MetricsMain.dox
#	doc/pages/MetricsPerformance.dox
#	doc/pages/Outdated.txt
#	doc/pages/a_PageOrder.dox
#	doc/setup/Doxyfile

+++ chassis/core/src/main/java/com/griddynamics/jagger/user/test/configurations/load/JLoadProfileUsers.java
@@ -6,14 +6,33 @@ import java.util.Objects;
 
 
 /**
- * This class represents a user group in {@link JLoadProfileUserGroups}.
+ * This class represents a group of virtual users.
+ * Available attributes:<p>
+ *     - numberOfUsers - A goal number of threads.<p>
+ *     - lifeTimeInSeconds - Describes how long threads will be alive. Default is 2 days.<p>
+ *     - startDelayInSeconds - Delay before first thread will start. Default is 0.<p>
+ *     - slewRateUsersPerSecond - Describes how many threads to start during every iteration. Default is numberOfUsers value.<p>
+ *
+ * Examples: @n
+ * @code
+ * JLoadProfileUsers u = JLoadProfileUsers.builder(NumberOfUsers.of(5)).withStartDelayInSeconds(10).build();
+ * @endcode
+ * @image html load_GroupWithDelay.png "User group load with start delay"
+ * @n
+ * @code
+ * JLoadProfileUsers u = JLoadProfileUsers.builder(NumberOfUsers.of(20)).withSlewRateUsersPerSecond(1).build();
+ * @endcode
+ * @image html load_GroupSlewRate.png "User group load with slew rate"
+ *
+ * @ingroup Main_Load_profiles_group
+
  */
 public class JLoadProfileUsers {
 
     private final long numberOfUsers;
     private final long lifeTimeInSeconds;
     private final long startDelayInSeconds;
-    private final long slewRateUsersPerSecond;
+    private final double slewRateUsersPerSecond;
 
     private JLoadProfileUsers(Builder builder) {
         this.numberOfUsers = builder.numberOfUsers.value();
@@ -36,7 +55,7 @@ public class JLoadProfileUsers {
         private final NumberOfUsers numberOfUsers;
         private long lifeTimeInSeconds;
         private long startDelayInSeconds;
-        private long slewRateUsersPerSecond;
+        private double slewRateUsersPerSecond;
 
         /** Builder of the JLoadProfileUsers
          * @n
@@ -87,7 +106,7 @@ public class JLoadProfileUsers {
          * Optional: Slew rate users per second. Default is numberOfUsers value.
          * @param slewRateUsersPerSecond Describes how many threads to start during every iteration
          */
-        public Builder withSlewRateUsersPerSecond(long slewRateUsersPerSecond) {
+        public Builder withSlewRateUsersPerSecond(double slewRateUsersPerSecond) {
             if (slewRateUsersPerSecond <= 0) {
                 throw new IllegalArgumentException(
                         String.format("Slew rate users per second must be > 0. Provided value is %s", slewRateUsersPerSecond));
@@ -109,7 +128,7 @@ public class JLoadProfileUsers {
         return startDelayInSeconds;
     }
 
-    public long getSlewRateUsersPerSecond() {
+    public double getSlewRateUsersPerSecond() {
         return slewRateUsersPerSecond;
     }
 }
